backend/requirements.txt
asgiref==3.2.7
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
decorator==4.4.2
Django==3.0.6
django-cors-headers==3.2.1
djangorestframework==3.11.0
future==0.18.2
yandex-geocoder
idna==2.9
Pillow==7.1.2
psycopg2==2.8.5
pytz==2020.1
ratelim==0.1.6
requests==2.23.0
six==1.14.0
sqlparse==0.3.1
urllib3==1.25.9
